write right_triangle.rb

+++ hw_1/area_triangle.rb
@@ -1,7 +1,7 @@
 puts "Какая длина основания треугольника?"
-base = gets.chomp.to_i
+base = gets.chomp.to_f
 puts "Какая высота треугольника?"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 area =  (base * height) / 2
